Added orio firmware for SPI link testing Moved the main firmware into a seperate subdirectory to keep things organized.
The SPI link could be used as a "semi-reliable" transmit medium where
we discard packets that don't match checksums, but we don't retransmit
since the rate of occurance is so low and we assume that the messages
are mostly indempotent. To do this we have to verify the SPI link
reliability.

Note that handling the packets this way means that non-indempotent
updates like joystick inputs will need to have some amount of ack
handling for things like button inputs or the like.